requirements.txt
picamera2
opencv-python
torch
torchvision
pillow
pandas
requests
